[lua] Removed animated_weapon.lua and all references to the mixin
animated_weapon.lua was an outdated relic of the 75 cap Dynamis days where the killing of the 15 demon NMs on the 5 towers in Dynamis-Xarcabard was required before the Animated Weapons dropped their relevant fragment. The mixin, along with the previously removed function activateAnimatedWeapon (removed in https://github.com/LandSandBoat/server/commit/56e5f4d70424a755968e4ed7fe99e046709f586a) are what allowed animation sub-state of the the Animated Weapons to drop their fragment.

Neo Dynamis now only requires the relevant goad item to be obtained to spawn the relevant animated weapon. Once the relevant animated weapon is defeated, the relevant fragment is a guaranteed 100% drop.

Reference the BG-Wiki pages for the Animated Weapons to note that all list the relevant fragment as a Guaranteed drop off the animated weapon.